specific_job/api/requirements.txt
# FastAPI packages:
uvicorn
fastapi
pydantic
pydantic_settings
python-multipart

# Services packages:
pandas==2.0.3
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
gunicorn==21.2.0